// src/app/shared/models/dataTableModel.ts
export class DataTableModel {

    // constructor(id: number, name: string, firstname: string, type: string){
    //     this.id = id;
    //     this.name = name;
    //     this.firstname = firstname;
    //     this.type = type;
    // }

    id: number;
    name: string;
    firstname: string;
    type: string;
}